2024-04-04-build-rag-with-python/requirements.txt
ocrmypdf==16.1.2
ollama==0.1.8
python_magic==0.4.27
Requests==2.31.0
